Unvendor flask-session
